GH-110 Add feedback that compares to test average
